Support for local, testing, and production env
The api mode can be changed via the `GAT.apiMode` global which
determines what api to connect to: a local, testing, or production.

Removed code that asked the api what url to connect to for socket.io.
This is no longer needed due to changes in the backend so now we
can connect directly to the notification/socket.io service without
asking the api.